requirements.txt
# Agno AI Agent Framework Dependencies
agno>=2.2.0

# Configuration
pyyaml>=6.0

# Memory Management
mem0ai>=1.0.0
fastapi[standard]
mcp

# Core AI Model Dependencies
openai>=2.0.0
google-generativeai>=0.8.0
google-genai>=1.50.0
anthropic>=0.7.0
ollama>=0.6.0

# Core Dependencies
pydantic>=2.9.0
httpx>=0.28.0
requests>=2.31.0
# Optional but recommended for enhanced functionality
# Uncomment as needed:

# For web scraping and API calls
# beautifulsoup4>=4.12.0

# For data processing
# pandas>=2.0.0
# numpy>=1.24.0

# For development and testing
# pytest>=7.0.0
# pytest-asyncio>=0.21.0